Add python functional test
This can be run by using `python testsuite_concordance.py`.
It currently still contains a number of errors (mainly
because page load times are too high). In the future it could be
used to time page load times. Also it would be beneficial to add
a way of making a test fail if the page load time is too high.